Move actions to a separate file
